% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEKae.R
\name{calculateEKae}
\alias{calculateEKae}
\title{calculateEKae function}
\usage{
calculateEKae(geno_data, A, E, site_env_qtl_all, m_env_qtl)
}
\arguments{
\item{geno_data}{genotype data}

\item{A}{additive relationship matrix}

\item{E}{environment variance-covariance matrix}

\item{site_env_qtl_all}{a dataframe with all AE QTL}

\item{m_env_qtl}{the number of AE QTL}
}
\value{
a list of KaeE1 and KaeE2
}
\description{
calculateEKae function
}
\examples{
\dontrun{EKae = calculateEKae(geno_data = mmgeno_data, A, E, site_env_qtl_all, m_env_qtl)}
}
