% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateKaXa.R
\name{calculateKaXa}
\alias{calculateKaXa}
\title{calculateKaXa function}
\usage{
calculateKaXa(geno_data, pheno_data, A, site_qtl, m_qtl)
}
\arguments{
\item{geno_data}{genotype data}

\item{pheno_data}{phenotype data}

\item{A}{additive relationship matrix}

\item{site_qtl}{a vector, names of QTL sites}

\item{m_qtl}{the number of QTLs}
}
\value{
list(Ka = Ka,Xa = Xa)
}
\description{
calculateKaXa function
}
\examples{
\dontrun{KaXa = calculateKaXa(mmgeno_data, mmpheno_data, A, site_qtl, m_qtl)}
}
