% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gblup.R
\name{gblup}
\alias{gblup}
\title{gblup function}
\usage{
gblup(data, A)
}
\arguments{
\item{data}{a data frame}

\item{A}{additive genetic relationship matrix}
}
\value{
list(mod, BV)
}
\description{
gblup function
}
\examples{
\dontrun{rst = gblup(data = dt, A = mmdata$A)}
}
