% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geblup.R
\name{geblup}
\alias{geblup}
\title{geblup function}
\usage{
geblup(data, A, AE)
}
\arguments{
\item{data}{a data frame}

\item{A}{additive genetic relationship matrix}

\item{AE}{additive-by-environment relationship matrix1}
}
\value{
list(mod, BV)
}
\description{
geblup function
}
\examples{
\dontrun{rst = geblup(data = dt, A = mmdata$A, AE = mmdata$AE)}
}
