% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgeblup.R
\name{mmgeblup}
\alias{mmgeblup}
\title{mmgeblup function}
\usage{
mmgeblup(data, Ka, EKae1, EKae2)
}
\arguments{
\item{data}{a data frame}

\item{Ka}{additive genetic relationship matrix}

\item{EKae1}{additive-by-environment relationship matrix1}

\item{EKae2}{additive-by-environment relationship matrix2}
}
\value{
list(mod, BV)
}
\description{
mmgeblup function
}
\examples{
\dontrun{rst = mmgeblup(data = cbind(dt, mmdata$Xa), Ka = mmdata$Ka, EKae1 = mmdata$EKae1, EKae2 = mmdata$EKae2)}
}
