% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtxnetwork.input.trans.R
\name{qtxnetwork.input.trans}
\alias{qtxnetwork.input.trans}
\title{qtxnetwork.input.trans function}
\usage{
qtxnetwork.input.trans(
  geno_data,
  pheno_data,
  geno_output_prefix,
  pheno_output_prefix
)
}
\arguments{
\item{geno_data}{genotype data, output from geno.generate()}

\item{pheno_data}{phentype data, output from pheno.generate()}

\item{geno_output_prefix}{output prefix for .gen file}

\item{pheno_output_prefix}{output prefix for .phe file}
}
\description{
qtxnetwork.input.trans function
}
\examples{
\dontrun{qtxnetwork.input.trans(geno_data, pheno_data,
                                          geno_output_prefix = geno_output_prefix,
                                          pheno_output_prefix = pheno_output_prefix)}
}
