% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtxnetwork.perform.R
\name{qtxnetwork.perform}
\alias{qtxnetwork.perform}
\title{qtxnetwork.perform function
1D scanning for additive and additive-by-environment qtls}
\usage{
qtxnetwork.perform(qtxnetwork_path, gen_file, phe_file, pre_file)
}
\arguments{
\item{qtxnetwork_path}{path to the executable QTXNetwork file}

\item{gen_file}{(input file) path and name for .gen file}

\item{phe_file}{(input file) path and name for .phe file}

\item{pre_file}{(output file) path and name for .pre file}
}
\description{
qtxnetwork.perform function
1D scanning for additive and additive-by-environment qtls
}
\examples{
\dontrun{qtxnetwork.perform(
                   qtxnetwork_path = "/public3/zqx/QTXNetwork_4.0/build/QTXNetwork"
                   gen_file = "./inst/Simulation.gen"
                   phe_file = "./inst/Simulation_SimTrait.phe"
                   pre_file = "./inst/Simulation_SimTrait.pre")}
}
