% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.effect.plot.R
\name{snp.effect.plot}
\alias{snp.effect.plot}
\title{snp.effect.plot function}
\usage{
snp.effect.plot(effects)
}
\arguments{
\item{effects}{an envNum * snpNum matrix for total SNP effect}
}
\value{
a ggplot object
}
\description{
snp.effect.plot function
}
\examples{
\dontrun{p = snp.effect.plot(effects = b)}
}
