% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateEKae.R
\name{calculateEKae}
\alias{calculateEKae}
\title{calculateEKae function}
\usage{
calculateEKae(geno_data, E, site_env_qtl_all)
}
\arguments{
\item{geno_data}{genotype data}

\item{E}{the environment covariance matrix}

\item{site_env_qtl_all}{a dataframe with all AE QTL}
}
\value{
a list of KaeE1 and KaeE2
}
\description{
calculateEKae function
}
\examples{
\dontrun{EKae = calculateEKae(geno_data = mmgeno_data, E, site_env_qtl_all)}
}
