% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmgeblup.R
\name{mmgeblup}
\alias{mmgeblup}
\title{mmgeblup function}
\usage{
mmgeblup(data, Ka, KaeE1, KaeE2)
}
\arguments{
\item{data}{a data frame}

\item{Ka}{additive genetic relationship matrix}

\item{KaeE1}{additive-by-environment relationship matrix1}

\item{KaeE2}{additive-by-environment relationship matrix2}
}
\value{
list(mod, BV)
}
\description{
mmgeblup function
}
\examples{
\dontrun{rst = mmgeblup(data = cbind(dt, mmdata$Xa), Ka = mmdata$Ka, KaeE1 = mmdata$KaeE1, KaeE2 = mmdata$KaeE2)}
}
